Reconfigure DB model to split user types. Reconfigure login, notification, and delete user functions.

+++ chornado_app/routes.py
@@ -1,10 +1,10 @@
-from flask import redirect, render_template, request, url_for, Blueprint
+from flask import redirect, render_template, request, url_for, Blueprint, flash
 from flask_login import current_user, login_required, logout_user
 from werkzeug.security import generate_password_hash
 
 from .forms import (DeleteUserForm, ChildResetPasswordForm)
-from .sql_models import (db, User, Notification)
-from .helpers import (db_commit, flash_errors)
+from .sql_models import (db, Parent, Child, ParentNotification)
+from .helpers import (db_commit, flash_errors, redirect_url)
 
 routes_bp = Blueprint('routes', __name__)
 
@@ -18,8 +18,8 @@ def index():
     # If logged in user is type: parent, load parent homepage
     if current_user.type == 'parent':
         return redirect(url_for('parent.home'))
-    # If logged in user is type: child, load child homepage
 
+    # If logged in user is type: child, load child homepage
     return redirect(url_for('child.home'))
 
 @routes_bp.route('/delete/<int:user_id>', methods=['GET', 'POST'])
@@ -27,20 +27,26 @@ def index():
 def delete_user(user_id):
     """Route for deleting users"""
 
-    user = User.query.get(user_id)
+    print(redirect_url())
+    print(url_for('parent.children'))
+    print(request.referrer)
+    if url_for('parent.children') in redirect_url():
+        print('child')
+        user = Child.query.get(user_id)      
+        
+    elif url_for('parent.settings') in redirect_url():
+        user = Parent.query.get(user_id)
+    # else:
+    #     user = None
     form = DeleteUserForm()
 
     if request.method == 'POST' and form.validate_on_submit():
-        if user.type == 'parent':
-            # Identify child accounts linked to parent account for deletion
-            children = User.query.filter_by(parent=user.id).all()
-            for child in children:
-                db.session.delete(child)
+        if user.type == 'parent' and user is not None:
             logout_user()
             db.session.delete(user)
         else:
             # Identify notifications linked to child account for deletion
-            notifications = Notification.query.filter_by(child_id=user.id).all()
+            notifications = ParentNotification.query.filter_by(child_id=user.id).all()
             for notification in notifications:
                 db.session.delete(notification)
             db.session.delete(user)
@@ -48,8 +54,53 @@ def delete_user(user_id):
         db_commit()
         return redirect(url_for('routes.index'))
 
-    if user is not None and (current_user.id in (user.parent, user_id)):
+    if user is not None and (user == current_user or user in current_user.children):
         return render_template('delete_user.html', template_form=form,
+            user=user)
+
+    return redirect(url_for('routes.index'))
+
+@routes_bp.route('/delete_child/<int:user_id>', methods=['GET', 'POST'])
+@login_required
+def delete_child(user_id):
+    """Route for deleting child users"""
+
+    user = Child.query.get(user_id)
+    form = DeleteUserForm()
+
+    if request.method == 'POST' and form.validate_on_submit():
+        # Identify notifications linked to child account for deletion
+        notifications = ParentNotification.query.filter_by(child_id=user.id).all()
+        for notification in notifications:
+            db.session.delete(notification)
+        flash(f'{user.first_name} has been deleted.', 'success')
+        db.session.delete(user)
+        db_commit()
+        return redirect(url_for('routes.index'))
+
+    if user is not None and user in current_user.children:
+        return render_template('delete_child.html', template_form=form,
+            user=user)
+    
+    return redirect(url_for('routes.index'))
+        
+@routes_bp.route('/delete_parent/<int:user_id>', methods=['GET', 'POST'])
+@login_required
+def delete_parent(user_id):
+    """Route for deleting parent users"""
+
+    user = Parent.query.get(user_id)
+    form = DeleteUserForm()
+
+    if request.method == 'POST' and form.validate_on_submit():
+        flash(f'{user.first_name} has been deleted.', 'success')
+        logout_user()
+        db.session.delete(user)
+        db_commit()
+        return redirect(url_for('routes.index'))
+
+    if user is not None and user == current_user:
+        return render_template('delete_parent.html', template_form=form,
             user=user)
 
     return redirect(url_for('routes.index'))
@@ -59,7 +110,7 @@ def delete_user(user_id):
 def pass_reset(user_id):
     """Route for resetting child passwords"""
 
-    user = User.query.get(user_id)
+    user = Child.query.get(user_id)
     form = ChildResetPasswordForm()
 
     if request.method == 'POST' and form.validate_on_submit():
@@ -69,7 +120,7 @@ def pass_reset(user_id):
         db_commit()
         return redirect(url_for('parent.children'))
 
-    if user is not None and user.parent == current_user.id:
+    if user is not None and user.parent_id == current_user.id:
         flash_errors(form)
         return render_template('pass_reset.html', template_form=form, user=user)
 
@@ -78,8 +129,9 @@ def pass_reset(user_id):
 # Route for testing database inserts and modifications
 @routes_bp.route('/test')
 def test():
-#     notifications = current_user.notifications
-#     for notification in notifications:
-#         db.session.delete(notification)
-#     db_commit()
+    # from sqlalchemy import exc
+    # try:
+    #     db.create_all()
+    # except exc.SQLAlchemyError as error:
+    #     print(error)
     return redirect(url_for('routes.index'))
